+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   minimum supported compiler version when using the `nightly` feature to 1.49.
 - Rename `Array{Vec, Deque, Heap}` to `Inline*` for consistency with `InlineObject`
 - Redefine `ArenaStorage` as a struct for compatibility with non-array-like layouts
+- Remove `HeapStorage` type alias and add `AllocStorage` struct (similar to `ArenaStorage`)
 
 [storage-abstraction-v2]: https://gist.github.com/teryror/7b9a23fd0cd8dcfbcb6ebd34ee2639f8